🎨  create new review fetching logic to server-side utility function to avoid cache problems
